feat:#157 header의 profile 수정
